==> 1_brms_binomial.R <==
library(data.table)
library(ape)
library(brms)

prepare_data <- F

FULL_MODELS <- "DONE"
FULL_MODELS_SAMPLE <- "RUN"
MODELS_ONE <- "DONE"
MODELS_MOLECULAR <- "DONE"

cluster_dir <- here::here("Data_for_cluster")
model_dir <- here::here("Models")

if(grepl("nbogdanovic", getwd())){
  
  main_dir <- "/home/nbogdanovic/WNV_prevalence"
  cluster_dir <- file.path(main_dir, "Data_for_cluster")
  model_dir <- file.path(main_dir, "Models")
  dir.create(model_dir, showWarnings = F)

  }

# Seed stuff
set.seed(202510)
BAYES_SEED <- 202510



# 0: Prepare data PC---------------------------------------------------------

if(prepare_data == T){
  
  # # laptop
  data_dir <- here::here("Data")
  # data for cluster dir
  cluster_dir <- here::here("Data_for_cluster")
  dir.create(cluster_dir, showWarnings = F)
  
  wnv_dt <- fread(file.path(data_dir, "00_WNV_prevalence_data.csv"))[
    , name_type := NULL]
  
  wnv_dt <- wnv_dt[
    , ":=" (
      avilist_name = gsub(" ", "_", avilist_name),
      method = gsub(" ", "_", method)
    )]
  
  # selected traits
  # removed "habitat_TP2019",
  traits_select <- c(
    "birdlife_name", "mass_log", "tarsus_log", "hwi_log", "longevity_log",
    "clutch_max_log", "log_human_population_density", "abundance_log",
    "habitat", "freshwater", "migration", "sociality",
    "primary_lifestyle", "nest_placement", "trophic_niche", "altitude_cat"
  )
  
  traits_dt <- fread(file.path(data_dir, "00_traits_filtered.csv"))
  
  traits_dt <- traits_dt[
    , ':=' (
      mass_log = scale(log(mass)),
      tarsus_log = scale(log(tarsus_length)),
      hwi_log = scale(log(hand_wing_index)),
      longevity_log = scale(log(maximum_longevity)),
      clutch_max_log = scale(log(clutch_max)),
      abundance_log = scale(log(abundance_estimate + 1)),
      log_human_population_density = scale(log_human_population_density),
      migration = factor(
        fifelse(migration == "sedentary", "sedentary", "migratory")),
      nest_placement = factor(nest_placement),
      habitat = fcase(
        habitat == "Human Modified", "humanModified",
        habitat %in% c("Forest", "Woodland", "Shrubland", "Grassland"), "greenLandscape",
        habitat %in% c("Wetland", "Riverine", "Coastal"), "waterLandscape",
        habitat %in% c("Desert", "Rock"), "rockyLandscape",
        habitat == "Marine", "marineLandscape"
      ),
      freshwater = fifelse(freshwater == 1, "freshwater", "other"),
      sociality = fcase(
        colonial == 1 | social == 1, "social",
        default = "nonSocial"),
      primary_lifestyle = fcase(
        primary_lifestyle %in% c("Aerial", "Insessorial"), "air",
        primary_lifestyle == "Aquatic", "water",
        default = "other"
      ),
      trophic_niche = fcase(
        trophic_niche %in% c("Scavenger", "Vertivore"), "carnivore",
        trophic_niche %in% c("Aquatic predator",  "Herbivore aquatic"), "aquaticFeeder",
        trophic_niche %in% c("Frugivore", "Granivore", "Nectarivore"), "cropFeeder",
        default = "terrestrialFeeder"),
      altitude_cat = fcase(
        minimum_altitude < 500, "lowland",
        minimum_altitude >= 500 & minimum_altitude < 1500, "midland",
        minimum_altitude >= 1500, "highland"
      )
    )][, ..traits_select]
  
  wnv_dt <- merge(wnv_dt, traits_dt, by = "birdlife_name")
  
  
  # keep only the data that is complete
  wnv_dt[
    , data_complete := apply(.SD, 1, function(row) all(!is.na(row) & row != ""))]
  wnv_dt <- wnv_dt[data_complete == T][, data_complete := NULL]
  
  # and minimum 20 tested individuals per species
  wnv_dt[, ind_per_sp := sum(total_tested), by = avilist_name]
  wnv_dt <- wnv_dt[ind_per_sp >= 20]
  
  big_tests <- c("serological_ELISA", "molecular_detection", "serological_VNT")
  wnv_dt[, method_cat := fifelse(!method %in% big_tests, "other", method)]
  
  wnv_dt[, ':=' (
    freshwater = relevel(factor(freshwater), ref = "other"),
    migration = relevel(factor(migration), ref = "sedentary"),
    primary_lifestyle = relevel(factor(primary_lifestyle), ref = "other"),
    trophic_niche = relevel(factor(trophic_niche), ref = "terrestrialFeeder"),
    altitude_cat = relevel(factor(altitude_cat), ref = "lowland"),
    habitat = relevel(factor(habitat), ref = "greenLandscape"),
    sociality = relevel(factor(sociality), ref = "nonSocial")
  )]
  
  #PHYLOGENETIC AUTOCORRELATION
  bird_tree <- readRDS(file.path(data_dir, "00_bird_tree_match_avilist.rds"))
  bird_tree <- drop.tip(
    bird_tree,
    bird_tree$tip.label[!bird_tree$tip.label %in% unique(wnv_dt$avilist_name)]
    )
  # this way we generate covariance matrix that will be inputted in the model as:
  # gr(scientific_name_bt, cov = A) - we make sure that species are correlated
  # as specified by the covariance matrix A
  A <- vcv.phylo(bird_tree)
  saveRDS(A, file.path(cluster_dir, "1_bird_tree_A.rds"))
  
  # locational autocorrelation
  study_locs <- unique(wnv_dt[, .(country, long, lat)])
  setorder(study_locs, country, long, lat)
  study_locs[
    , location_id := paste0(gsub(" ", "_", country), "_", 1:.N), by = country]
  saveRDS(study_locs, file.path(cluster_dir, "1_study_locations.rds"))
  
  # study_locs <- sf::st_as_sf(study_locs, coords = c("long", "lat"), crs = 4326)
  # spatiald <- distm(sf::st_distance(study_locs))/1000 # in km
  
  # get distance matrix and convert to kilometers
  dist_matrix_kmAF <- geosphere::distm(
    x = study_locs[, c("long", "lat")], fun = geosphere::distHaversine)/1000 # in M
  # in KM
  
  # Rename matrix with location_id
  rownames(dist_matrix_kmAF) <- study_locs$location_id
  colnames(dist_matrix_kmAF) <- study_locs$location_id
  range_paramAF <- 500  # decay scale in km
  Sigma_spatialAF <- exp(-dist_matrix_kmAF / range_paramAF)
  
  saveRDS(
    Sigma_spatialAF, file.path(cluster_dir, "1_spatial_autocorrelation.rds"))
  
  
  wnv_dt <- merge(
    wnv_dt, study_locs[, .(location_id, country, long, lat)], 
    by = c("country", "long", "lat")
  )
  
  # save data
  saveRDS(wnv_dt, file.path(cluster_dir, "1_WNV_prevalence_data_model.rds"))
  
}


# add_continents <- function(
#     steps, crs = sf::st_crs(4326), align_start = T,
#     coord_cols = NULL, scale = "medium"){
#
#   # overlay the map with start or the end point of the step (didn't check both
#   # to preserve the country information in case we need it later)
#   if(is.null(coord_cols)){
#     coord_cols <- if(align_start) c("x1_", "y1_") else c("x2_", "y2_")
#   }
#
#   steps <- steps[
#     , geometry := sf::st_as_sf(
#       .SD, coords = coord_cols, crs = crs), .SDcols = coord_cols]
#
#   # load or download the continent data with the right resolution
#   conti <- tryCatch(
#     rnaturalearth::ne_load(
#       type = "geography_regions_polys",  scale = scale, category = "physical"),
#     error = function(e) {
#       rnaturalearth::ne_download(
#         type = "geography_regions_polys",  scale = scale, category = "physical")
#     }
#   )
#
#   # check the nearest geometry
#   conti_id <- sf::st_nearest_feature(steps$geometry, conti)
#
#   # convert to datatable for merging
#   conti_dt <- as.data.table(conti)[
#     , .(continent = REGION, subregion = SUBREGION)]
#   conti_dt <- conti_dt[conti_id]
#
#   # add columns to the original step data
#   steps <- cbind(steps[, geometry := NULL], conti_dt)
#
#   return(steps)
#
# }
#
#
# dist_locs <- add_continents(
#   wnv_dt, coord_cols = c("long", "lat"), scale = "medium")


# 0: Prepare subset data --------------------------------------------------

if(prepare_data == T){

  # Load data
  wnv_dt <- readRDS(file.path(cluster_dir, "1_WNV_prevalence_data_model.rds"))
  bird_tree <- readRDS(file.path(data_dir, "00_bird_tree_match_avilist.rds"))
  study_locs <- readRDS(file.path(cluster_dir, "1_study_locations.rds"))
  
  ### ONE ENTRY PER SPECIES
  # select one entry per species 
  setorder(wnv_dt, avilist_name, -total_tested)
  wnv_dt_one <- wnv_dt[, .SD[1], by = avilist_name][total_tested >= 10]
  saveRDS(
    wnv_dt_one, file.path(cluster_dir, "1_WNV_prevalence_data_model_one.rds"))
  
  # prepare bird tree
  bird_tree_one <- drop.tip(
    bird_tree,
    bird_tree$tip.label[!bird_tree$tip.label %in% unique(wnv_dt_one$avilist_name)]
  )
  A_one <- vcv.phylo(bird_tree_one)
  saveRDS(A_one, file.path(cluster_dir, "1_bird_tree_A_one.rds"))
  
  # get distance matrix and convert to kilometers
  study_locs_one <- study_locs[location_id %in% unique(wnv_dt_one$location_id)]
  dist_matrix_kmAF_one <- geosphere::distm(
    x = study_locs_one[, c("long", "lat")], 
    fun = geosphere::distHaversine)/1000 # in KM
  
  # Rename matrix with location_id
  rownames(dist_matrix_kmAF_one) <- study_locs_one$location_id
  colnames(dist_matrix_kmAF_one) <- study_locs_one$location_id
  range_paramAF <- 500  # decay scale in km
  Sigma_spatialAF_one <- exp(-dist_matrix_kmAF_one / range_paramAF)
  
  saveRDS(
    Sigma_spatialAF_one, 
    file.path(cluster_dir, "1_spatial_autocorrelation_one.rds")
  )
  rm(wnv_dt_one, dist_matrix_kmAF_one, study_locs_one, 
     bird_tree_one, A_one, Sigma_spatialAF_one)
  
  ### ONLY ONE METHOD
  wnv_dt_molecular <- wnv_dt[method == "molecular_detection"]
  saveRDS(
    wnv_dt_molecular, 
    file.path(cluster_dir, "1_WNV_prevalence_data_model_molecular.rds")
  )
  
  # prepare bird tree
  bird_tree_molecular <- drop.tip(
    bird_tree,
    bird_tree$tip.label[!bird_tree$tip.label %in% unique(wnv_dt_molecular$avilist_name)]
  )
  
  A_molecular <- vcv.phylo(bird_tree_molecular)
  saveRDS(A_molecular, file.path(cluster_dir, "1_bird_tree_A_molecular.rds"))
  
  # get distance matrix and convert to kilometers
  study_locs_molecular <- study_locs[location_id %in% unique(wnv_dt_molecular$location_id)]
  dist_matrix_kmAF_molecular <- geosphere::distm(
    x = study_locs_molecular[, c("long", "lat")], 
    fun = geosphere::distHaversine)/1000 # in KM
  
  # Rename matrix with location_id
  rownames(dist_matrix_kmAF_molecular) <- study_locs_molecular$location_id
  colnames(dist_matrix_kmAF_molecular) <- study_locs_molecular$location_id
  range_paramAF <- 500  # decay scale in km
  Sigma_spatialAF_molecular <- exp(-dist_matrix_kmAF_molecular / range_paramAF)
  
  saveRDS(
    Sigma_spatialAF_molecular, 
    file.path(cluster_dir, "1_spatial_autocorrelation_molecular.rds")
  )
  rm(wnv_dt_molecular, dist_matrix_kmAF_molecular, study_locs_molecular,
     bird_tree_molecular, A_molecular, Sigma_spatialAF_molecular)
  
  ### SAMPLE SIZE PER STUDY
  wnv_dt_sample <- wnv_dt[total_tested >= 10]
  saveRDS(
    wnv_dt_sample, file.path(cluster_dir, "1_WNV_prevalence_data_model_sample.rds"))
  
  # prepare bird tree
  bird_tree_sample <- drop.tip(
    bird_tree,
    bird_tree$tip.label[!bird_tree$tip.label %in% unique(wnv_dt_sample$avilist_name)]
  )
  A_sample <- vcv.phylo(bird_tree_sample)
  saveRDS(A_sample, file.path(cluster_dir, "1_bird_tree_A_sample.rds"))
  
  # get distance matrix and convert to kilometers
  study_locs_sample <- study_locs[location_id %in% unique(wnv_dt_sample$location_id)]
  dist_matrix_kmAF_sample <- geosphere::distm(
    x = study_locs_sample[, c("long", "lat")], 
    fun = geosphere::distHaversine)/1000 # in KM
  
  # Rename matrix with location_id
  rownames(dist_matrix_kmAF_sample) <- study_locs_sample$location_id
  colnames(dist_matrix_kmAF_sample) <- study_locs_sample$location_id
  range_paramAF <- 500  # decay scale in km
  Sigma_spatialAF_sample <- exp(-dist_matrix_kmAF_sample / range_paramAF)
  
  saveRDS(
    Sigma_spatialAF_sample, 
    file.path(cluster_dir, "1_spatial_autocorrelation_sample.rds")
  )
  rm(wnv_dt_sample, dist_matrix_kmAF_sample, study_locs_sample, 
     bird_tree_sample, A_sample, Sigma_spatialAF_sample)
}


# 1: FULL_MODELS ---------------------------------------------------------------

# checking which variables are correlated
# log_var <- grep("^log_|_log$", names(wnv_dt), value = T)
# 
# psych::pairs.panels(wnv_dt[, ..log_var],
#              method = "pearson",
#              hist.col = "#00AFBB",
#              density = TRUE)

if(FULL_MODELS == "RUN"){
  
  
  A <- readRDS(file.path(cluster_dir, "1_bird_tree_A.rds"))
  Sigma_spatialAF <- readRDS(
    file.path(cluster_dir, "1_spatial_autocorrelation.rds"))
  wnv_dt <- readRDS(file.path(cluster_dir, "1_WNV_prevalence_data_model.rds"))
  wnv_dt[, avilist_species := avilist_name]
  setwd(model_dir)
  
  
  full_f <- "positive | trials(total_tested) ~ 1 + tarsus_log + longevity_log + clutch_max_log + log_human_population_density + abundance_log + freshwater + migration + sociality + primary_lifestyle + nest_placement + trophic_niche + (1 | gr(avilist_name,cov=A)) + (1 | avilist_species) + (1 | method_cat) + (1 | gr(location_id, cov = Sigma_spatialAF)) + (1 | sampling_year)"
  
  
  # get the log warning files
  logfile <- file.path(model_dir, "1_check_distributions_warnings.log")
  logcon <- file(logfile, open = "at")
  
  distributions <- list(
    binom = binomial(), 
    beta_binom = beta_binomial(), 
    zero_binom = zero_inflated_binomial(), 
    zero_beta_binom = zero_inflated_beta_binomial()
  )
  
  for(dn in names(distributions)){
    
    # TRY
    mname <- paste0("1_", dn, "_full_model_spatial_random_sp.rds")
    
    # Fit the model only if not already saved
    if (!file.exists(mname)) {
      
      # convert to brms formula object
      f <- bf(formula(full_f), family = distributions[[dn]])
      
      withCallingHandlers({
        m <- brm(
          data = wnv_dt,
          data2 = list(
            A = A, 
            Sigma_spatialAF = Sigma_spatialAF
          ),
          formula = f,
          iter = 6000,
          warmup = 2000,
          cores = 4,
          chains = 4,
          seed = BAYES_SEED,
          sample_prior = TRUE,
          file = mname
        )
      }, warning = function(w) {
        writeLines(
          paste(Sys.time(), " \n | MODEL:", mname, " \n", conditionMessage(w)),
          logcon
        )
        invokeRestart("muffleWarning")
      })
    }
    
    writeLines(c(" ", "MODEL:", mname, "DONE!", " "))
    
  }
  
  
  close(logcon) # Close log file connection
  
  rm(wnv_dt, A, Sigma_spatialAF)
  
  
}


# 1: FULL_MODELS_SAMPLE -----------------------------------------------------

if(FULL_MODELS_SAMPLE == "RUN"){
  
  
  A <- readRDS(file.path(cluster_dir, "1_bird_tree_A_sample.rds"))
  Sigma_spatialAF <- readRDS(
    file.path(cluster_dir, "1_spatial_autocorrelation_sample.rds"))
  wnv_dt <- readRDS(file.path(cluster_dir, "1_WNV_prevalence_data_model_sample.rds"))
  wnv_dt[, avilist_species := avilist_name]
  setwd(model_dir)
  
  
  full_f <- "positive | trials(total_tested) ~ 1 + tarsus_log + longevity_log + clutch_max_log + log_human_population_density + abundance_log + freshwater + migration + sociality + primary_lifestyle + nest_placement + trophic_niche + (1 | gr(avilist_name,cov=A)) + (1 | avilist_species) + (1 | method_cat) + (1 | gr(location_id, cov = Sigma_spatialAF)) + (1 | sampling_year)"
  
  
  # get the log warning files
  logfile <- file.path(model_dir, "1_check_distributions_warnings.log")
  logcon <- file(logfile, open = "at")
  
  distributions <- list(
    binom = binomial(), 
    beta_binom = beta_binomial(), 
    zero_binom = zero_inflated_binomial(), 
    zero_beta_binom = zero_inflated_beta_binomial()
  )
  
  for(dn in names(distributions)){
    
    # TRY
    mname <- paste0("1_", dn, "_full_model_spatial_random_sp_sample.rds")
    
    # Fit the model only if not already saved
    if (!file.exists(mname)) {
      
      # convert to brms formula object
      f <- bf(formula(full_f), family = distributions[[dn]])
      
      withCallingHandlers({
        m <- brm(
          data = wnv_dt,
          data2 = list(
            A = A, 
            Sigma_spatialAF = Sigma_spatialAF
          ),
          formula = f,
          iter = 6000,
          warmup = 2000,
          cores = 4,
          chains = 4,
          seed = BAYES_SEED,
          sample_prior = TRUE,
          file = mname
        )
      }, warning = function(w) {
        writeLines(
          paste(Sys.time(), " \n | MODEL:", mname, " \n", conditionMessage(w)),
          logcon
        )
        invokeRestart("muffleWarning")
      })
    }
    
    writeLines(c(" ", "MODEL:", mname, "DONE!", " "))
    
  }
  
  
  close(logcon) # Close log file connection
  
  rm(wnv_dt, A, Sigma_spatialAF)
  
  
}





  
# 1: MODELS_ONE ----------------------------------------------------------------

if(MODELS_ONE == "RUN"){
  
  full_f <- "positive | trials(total_tested) ~ 1 + tarsus_log + longevity_log + clutch_max_log + log_human_population_density + abundance_log + freshwater + migration + sociality + primary_lifestyle + nest_placement + trophic_niche + (1 | gr(avilist_name,cov=A)) + (1 | method_cat) + (1 | gr(location_id, cov = Sigma_spatialAF)) + (1 | sampling_year)"

  
  A <- readRDS(file.path(cluster_dir, "1_bird_tree_A_one.rds"))
  Sigma_spatialAF <- readRDS(
    file.path(cluster_dir, "1_spatial_autocorrelation_one.rds"))
  wnv_dt <- readRDS(file.path(cluster_dir, "1_WNV_prevalence_data_model_one.rds"))
  setwd(model_dir)
  
  
  # get the log warning files
  logfile <- file.path(model_dir, "1_check_distributions_warnings.log")
  logcon <- file(logfile, open = "at")
  
  distributions <- list(
    binom = binomial(), 
    beta_binom = beta_binomial(), 
    zero_binom = zero_inflated_binomial(), 
    zero_beta_binom = zero_inflated_beta_binomial()
  )
  
  for(dn in names(distributions)){
    
    # TRY
    mname <- paste0("1_", dn, "_full_model_spatial_one.rds")
    
    # Fit the model only if not already saved
    if (!file.exists(mname)) {
      
      # convert to brms formula object
      f <- bf(formula(full_f), family = distributions[[dn]])
      
      withCallingHandlers({
        m <- brm(
          data = wnv_dt,
          data2 = list(
            A = A, 
            Sigma_spatialAF = Sigma_spatialAF
          ),
          formula = f,
          iter = 6000,
          warmup = 2000,
          cores = 4,
          chains = 4,
          seed = BAYES_SEED,
          sample_prior = TRUE,
          file = mname
        )
      }, warning = function(w) {
        writeLines(
          paste(Sys.time(), " \n | MODEL:", mname, " \n", conditionMessage(w)),
          logcon
        )
        invokeRestart("muffleWarning")
      })
    }
    
    writeLines(c(" ", "MODEL:", mname, "DONE!", " "))
    
  }
  
  
  close(logcon) # Close log file connection
  
  rm(wnv_dt, A, Sigma_spatialAF)
}


# 1: MODELS_MOLECULAR ----------------------------------------------------------

if(MODELS_MOLECULAR == "RUN"){
  
  full_f <- "positive | trials(total_tested) ~ 1 + tarsus_log + longevity_log + clutch_max_log + log_human_population_density + abundance_log + freshwater + migration + sociality + primary_lifestyle + nest_placement + trophic_niche + (1 | gr(avilist_name,cov=A)) + (1 | gr(location_id, cov = Sigma_spatialAF)) + (1 | sampling_year)"
  
  
  A <- readRDS(file.path(cluster_dir, "1_bird_tree_A_molecular.rds"))
  Sigma_spatialAF <- readRDS(
    file.path(cluster_dir, "1_spatial_autocorrelation_molecular.rds"))
  wnv_dt <- readRDS(
    file.path(cluster_dir, "1_WNV_prevalence_data_model_molecular.rds"))
  setwd(model_dir)
  
  
  # get the log warning files
  logfile <- file.path(model_dir, "1_check_distributions_warnings.log")
  logcon <- file(logfile, open = "at")
  
  distributions <- list(
    binom = binomial(), 
    beta_binom = beta_binomial(), 
    zero_binom = zero_inflated_binomial(), 
    zero_beta_binom = zero_inflated_beta_binomial()
  )
  
  for(dn in names(distributions)){
    
    # TRY
    mname <- paste0("1_", dn, "_full_model_spatial_molecular.rds")
    
    # Fit the model only if not already saved
    if (!file.exists(mname)) {
      
      # convert to brms formula object
      f <- bf(formula(full_f), family = distributions[[dn]])
      
      withCallingHandlers({
        m <- brm(
          data = wnv_dt,
          data2 = list(
            A = A, 
            Sigma_spatialAF = Sigma_spatialAF
          ),
          formula = f,
          iter = 6000,
          warmup = 2000,
          cores = 4,
          chains = 4,
          seed = BAYES_SEED,
          sample_prior = TRUE,
          file = mname
        )
      }, warning = function(w) {
        writeLines(
          paste(Sys.time(), " \n | MODEL:", mname, " \n", conditionMessage(w)),
          logcon
        )
        invokeRestart("muffleWarning")
      })
    }
    
    writeLines(c(" ", "MODEL:", mname, "DONE!", " "))
    
  }
  
  
  close(logcon) # Close log file connection
  
  rm(wnv_dt, A, Sigma_spatialAF)
  
}




# check models ------------------------------------------------------------

models_dir <- here::here("Models")


mnames <- list.files(models_dir, pattern = ".rds", full.names = T)

mone <- list.files(models_dir, pattern = "_one.rds", full.names = T)
one_models <- lapply(mone, readRDS)
names(one_models) <- basename(mone)
loo_one <- lapply(one_models[!grepl("1_beta_binom_full_", mone)], loo)
loo_compare(loo_one)


summary_one <- lapply(one_models, summary)


# # 
# # # Warning messages:
# # #   1: There were 7 divergent transitions after warmup. Increasing adapt_delta above 0.8 may help. See http://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup 
# # # 2: There were 13 divergent transitions after warmup. Increasing adapt_delta above 0.8 may help. See http://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup 
# # # 3: There were 6 divergent transitions after warmup. Increasing adapt_delta above 0.8 may help. See http://mc-stan.org/misc/warnings.html#divergent-transitions-after-warmup 
# # 
# loom <- lapply(all_models[3:6], loo)
# 
# loo_compare(loom)
# 
# summary(m)
# 
#
